Create README.md
# fetch_and_save_block_info
The Danish repository is intended for educational purposes. This repository shows how you can get the remaining hash of a block to find the hash of the next block.
To use the repository, you need to install two libraries, using the requests library, create a folder in which the Bitcoin block number and evo hash are stored in each text file.